chore: organization of components
