+++ apps/dapp/src/components/clamm/PositionsTable/components/Positions/components/columnHelpers/tradeHistory.tsx
